requirements.txt
streamlit
python-docx
PyMuPDF
pdfminer.six
language-tool-python
spacy
scikit-learn
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl